refactor: page progress bar by using useReading
