+++ apps/web/src/routes/kiwi/-reader/index.ts
@@ -1,7 +1,7 @@
 export * from "./book-context";
 export { default as ReaderContents } from "./contents";
 export * from "./flip-page-button";
-export * from "./page-progress";
+export { default as ReaderPageProgress } from "./page-progress";
 export { default as ReaderProvider } from "./provider";
 export * from "./settings-context";
 export * from "./styles";
